third_party/blink/web_tests/external/wpt/css/css-animations/responsive/column-width-001-expected.txt

This is a testharness.js-based test.
[FAIL] column-width clamps to 0px
  assert_equals: expected "20px" but got "0px"
Harness: the test ran to completion.
